BACKEND/Gestion_Ma_Ec/requirements.txt
Django==5.0.7
djangorestframework==3.17.1
django-cors-headers==4.8.0
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
requests==2.32.5
python-dotenv==1.0.1
